3 commit

+++ src/main/java/com/mycompany/tiendasaludable/VentaDetalle.java
@@ -46,7 +46,7 @@ public class VentaDetalle {
         this.dieta = dieta;
     }
 
-    public double getPrecio() {
+    public int getPrecio() {
         return precio;
     }
